consolidate and comment input_remapping.h
