+++ menu/cbs/menu_cbs_sublabel.c
@@ -21,6 +21,7 @@
 
 #include "../menu_driver.h"
 #include "../menu_cbs.h"
+#include "../../input/input_remapping.h"
 
 #include "../../retroarch.h"
 #include "../../core_option_manager.h"
