+++ ui/drivers/qt/qt_options.cpp
@@ -14,6 +14,7 @@ extern "C" {
 
 #include "../../../gfx/video_display_server.h"
 #include "../../../input/input_driver.h"
+#include "../../../input/input_remapping.h"
 #include "../../../network/netplay/netplay.h"
 #include "../../../retroarch.h"
